+++ src/screens/Vault/SigningDeviceList.tsx
@@ -3,13 +3,9 @@ import React, { useContext, useEffect, useState } from 'react';
 import { hp, windowHeight, windowWidth, wp } from 'src/common/data/responsiveness/responsive';
 
 import Alert from 'src/assets/images/alert_illustration.svg';
-import CVVInputsView from 'src/components/HealthCheck/CVVInputsView';
-import CustomGreenButton from 'src/components/CustomButton/CustomGreenButton';
-import DeleteIcon from 'src/assets/icons/deleteBlack.svg';
 import HardwareModalMap from './HardwareModalMap';
 import HeaderTitle from 'src/components/HeaderTitle';
 import KeeperModal from 'src/components/KeeperModal';
-import KeyPadView from 'src/components/AppNumPad/KeyPadView';
 import { LocalizationContext } from 'src/common/content/LocContext';
 import NFC from 'src/core/services/nfc';
 import { RFValue } from 'react-native-responsive-fontsize';
@@ -116,176 +112,6 @@ const SigningDeviceList = ({ navigation }: { navigation }) => {
     );
   };
 
-  const BulletPoint = ({ description }: { description: string }) => {
-    return (
-      <Box flexDirection={'row'} justifyContent={'center'} alignItems={'center'}>
-        <Box width={2} height={2} borderRadius={10} backgroundColor={'light.modalText'} />
-        <Text
-          fontSize={13}
-          fontWeight={200}
-          letterSpacing={0.65}
-          width={wp(260)}
-          color={'light.modalText'}
-          marginY={2}
-          marginLeft={3}
-        >
-          {description}
-        </Text>
-      </Box>
-    );
-  };
-
-  const settingSigningServer = () => {
-    return (
-      <Box>
-        {/* { this assert needs to be updated  } */}
-        <Alert />
-        <BulletPoint
-          description={
-            'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor'
-          }
-        />
-        <BulletPoint
-          description={
-            'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor'
-          }
-        />
-      </Box>
-    );
-  };
-
-  const setUpMobileKey = () => {
-    return (
-      <Box>
-        {/* { this assert needs to be updated  } */}
-        <Alert />
-        <BulletPoint
-          description={
-            'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor'
-          }
-        />
-        <BulletPoint
-          description={
-            'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor'
-          }
-        />
-      </Box>
-    );
-  };
-  const otpContent = () => {
-    const [otp, setOtp] = useState('');
-
-    console.log(otp);
-
-    const onPressNumber = (text) => {
-      let tmpPasscode = otp;
-      if (otp.length < 6) {
-        if (text != 'x') {
-          tmpPasscode += text;
-          setOtp(tmpPasscode);
-        }
-      }
-      if (otp && text == 'x') {
-        setOtp(otp.slice(0, -1));
-      }
-    };
-
-    const onDeletePressed = (text) => {
-      setOtp(otp.slice(0, otp.length - 1));
-    };
-
-    return (
-      <Box width={hp(280)}>
-        <Box>
-          <CVVInputsView
-            passCode={otp}
-            passcodeFlag={false}
-            backgroundColor={true}
-            textColor={true}
-          />
-          <Text
-            fontSize={13}
-            fontWeight={200}
-            letterSpacing={0.65}
-            width={wp(290)}
-            color={'light.modalText'}
-            marginTop={2}
-          >
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-            incididunt ut labore et
-          </Text>
-          <Box mt={10} alignSelf={'flex-end'} mr={2}>
-            <Box>
-              <CustomGreenButton onPress={() => {}} value={'proceed'} />
-            </Box>
-          </Box>
-        </Box>
-        <KeyPadView
-          onPressNumber={onPressNumber}
-          onDeletePressed={onDeletePressed}
-          keyColor={'light.lightBlack'}
-          ClearIcon={<DeleteIcon />}
-        />
-      </Box>
-    );
-  };
-  const passwordEnter = () => {
-    const [password, setPassword] = useState('');
-
-    console.log(password);
-
-    const onPressNumber = (text) => {
-      let tmpPasscode = password;
-      if (password.length < 6) {
-        if (text != 'x') {
-          tmpPasscode += text;
-          setPassword(tmpPasscode);
-        }
-      }
-      if (password && text == 'x') {
-        setPassword(password.slice(0, -1));
-      }
-    };
-
-    const onDeletePressed = (text) => {
-      setPassword(password.slice(0, password.length - 1));
-    };
-
-    return (
-      <Box width={hp(280)}>
-        <Box>
-          <CVVInputsView
-            passCode={password}
-            passcodeFlag={false}
-            backgroundColor={true}
-            textColor={true}
-          />
-          <Text
-            fontSize={13}
-            fontWeight={200}
-            letterSpacing={0.65}
-            width={wp(290)}
-            color={'light.modalText'}
-            marginTop={2}
-          >
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-            incididunt ut labore et
-          </Text>
-          <Box mt={10} alignSelf={'flex-end'} mr={2}>
-            <Box>
-              <CustomGreenButton onPress={() => {}} value={'proceed'} />
-            </Box>
-          </Box>
-        </Box>
-        <KeyPadView
-          onPressNumber={onPressNumber}
-          onDeletePressed={onDeletePressed}
-          keyColor={'light.lightBlack'}
-          ClearIcon={<DeleteIcon />}
-        />
-      </Box>
-    );
-  };
   return (
     <Box style={styles.container}>
       <StatusBarComponent padding={50} />
@@ -344,48 +170,6 @@ const SigningDeviceList = ({ navigation }: { navigation }) => {
           textColor={'#041513'}
           butt
           Content={nfcAlertConternt}
-        />
-        <KeeperModal
-          visible={false}
-          close={() => {}}
-          title={'Setting up a Signing Server'}
-          subTitle={'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed '}
-          modalBackground={['#F7F2EC', '#F7F2EC']}
-          buttonBackground={['#00836A', '#073E39']}
-          buttonText={'Continue'}
-          buttonTextColor={'#FAFAFA'}
-          textColor={'#041513'}
-          Content={settingSigningServer}
-        />
-        <KeeperModal
-          visible={false}
-          close={() => {}}
-          title={'Set up a Mobile Key'}
-          subTitle={'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed '}
-          modalBackground={['#F7F2EC', '#F7F2EC']}
-          buttonBackground={['#00836A', '#073E39']}
-          buttonText={'Continue'}
-          buttonTextColor={'#FAFAFA'}
-          textColor={'#041513'}
-          Content={setUpMobileKey}
-        />
-        <KeeperModal
-          visible={false}
-          close={() => {}}
-          title={'Confirm OTP to setup 2FA'}
-          subTitle={'Lorem ipsum dolor sit amet, '}
-          modalBackground={['#F7F2EC', '#F7F2EC']}
-          textColor={'#041513'}
-          Content={otpContent}
-        />
-        <KeeperModal
-          visible={false}
-          close={() => {}}
-          title={'Enter your password'}
-          subTitle={'Lorem ipsum dolor sit amet, '}
-          modalBackground={['#F7F2EC', '#F7F2EC']}
-          textColor={'#041513'}
-          Content={passwordEnter}
         />
       </Box>
     </Box>
